requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
mlxtend>=0.22
scipy>=1.10
xgboost>=2.0
pyyaml>=6.0
pyarrow>=12.0
jupyter>=1.0
ipykernel>=6.0
